refactor project tree and unified Storage methods
