+++ app/src/main/java/hpsaturn/pollutionreporter/view/ScanFragment.java
@@ -25,9 +25,8 @@ import java.util.concurrent.TimeUnit;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import hpsaturn.pollutionreporter.AppData;
-import hpsaturn.pollutionreporter.Keys;
+import hpsaturn.pollutionreporter.common.Keys;
 import hpsaturn.pollutionreporter.MainActivity;
-import hpsaturn.pollutionreporter.ScanResultsAdapter;
 import hpsaturn.pollutionreporter.R;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
