+++ app/src/main/java/hpsaturn/pollutionreporter/view/ScanResultsAdapter.java
@@ -1,4 +1,4 @@
-package hpsaturn.pollutionreporter;
+package hpsaturn.pollutionreporter.view;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -18,6 +18,7 @@ import java.util.Locale;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import hpsaturn.pollutionreporter.R;
 
 public class ScanResultsAdapter extends RecyclerView.Adapter<ScanResultsAdapter.ViewHolder> {
